+++ log_pytest/tests/conftest.py
@@ -1,0 +1,4 @@
+pytest_plugins = [
+    "devutils.fixtures",
+    "robocorp_log_tests.fixtures",
+]