+++ log_pytest/tests/robocorp_log_pytest_tests/test_robocorp_log_pytest_integrated/check_integration.py
@@ -1,0 +1,10 @@
+def test_ok():
+    a = 1  # noqa
+    print("worked")
+    for i in range(50):
+        j = "************" * i
+        print(j)
+
+
+def test_fail():
+    assert False, "Something failed..."