+++ tasks/tests/tasks_tests/test_load_autolog_config.py
@@ -1,6 +1,7 @@
 def test_load_autolog_config(tmpdir, str_regression) -> None:
     from pathlib import Path
-    from robocorp.tasks._toml_settings import read_pyproject_toml
+    from robocorp.log.pyproject_config import read_robocorp_auto_log_config
+    from robocorp.log.pyproject_config import read_pyproject_toml
 
     target = tmpdir / "pyproject.toml"
     target.write_text(
@@ -18,19 +19,19 @@ log_filter_rules = [
     pyproject_info = read_pyproject_toml(Path(target))
     assert pyproject_info is not None
 
-    from robocorp.tasks._log_auto_setup import read_robocorp_log_config
-
     class Ctx:
         def show_error(self, error):
             raise AssertionError(error)
 
-    config = read_robocorp_log_config(Ctx(), pyproject_info)
+    config = read_robocorp_auto_log_config(Ctx(), pyproject_info)
+
     str_regression.check(str(config))
 
 
 def test_load_autolog_config_exclude_default(tmpdir, str_regression) -> None:
     from pathlib import Path
-    from robocorp.tasks._toml_settings import read_pyproject_toml
+    from robocorp.log.pyproject_config import read_pyproject_toml
+    from robocorp.log.pyproject_config import read_robocorp_auto_log_config
 
     target = tmpdir / "pyproject.toml"
     target.write_text(
@@ -48,11 +49,9 @@ default_library_filter_kind = "exclude"
     pyproject_info = read_pyproject_toml(Path(target))
     assert pyproject_info is not None
 
-    from robocorp.tasks._log_auto_setup import read_robocorp_log_config
-
     class Ctx:
         def show_error(self, error):
             raise AssertionError(error)
 
-    config = read_robocorp_log_config(Ctx(), pyproject_info)
+    config = read_robocorp_auto_log_config(Ctx(), pyproject_info)
     str_regression.check(str(config))